Move c/h files implementing/defining standard library stuff into a new libc directory, also standard'ify some parts of the code base (almost entirely #include fixes). This is to a) to cleanup firmware/common and firmware/include a bit, but also b) for Rockbox as an application which should use the host system's c library and headers, separating makes it easy to exclude our files from the build.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@25850 a1c6a512-1295-4272-9138-f99709370657

+++ apps/bookmark.c
@@ -342,7 +342,7 @@ static char* create_bookmark()
              /* new optional bookmark token descriptors should be inserted
                 just before the "%s;%s" in this line... */
 #if CONFIG_CODEC == SWCODEC
-             ">%d;%d;%ld;%d;%ld;%d;%d;%d;%d;%s;%s",
+             ">%d;%d;%ld;%d;%ld;%d;%d;%ld;%ld;%s;%s",
 #else
              ">%d;%d;%ld;%d;%ld;%d;%d;%s;%s",
 #endif
@@ -360,8 +360,8 @@ static char* create_bookmark()
              global_settings.playlist_shuffle,
              /* ...and their values should go here */
 #if CONFIG_CODEC == SWCODEC
-             sound_get_pitch(),
-             dsp_get_timestretch(),
+             (long)sound_get_pitch(),
+             (long)dsp_get_timestretch(),
 #endif
              /* more mandatory tokens */
              playlist_get_name(NULL, global_temp_buffer,
